# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/v1/finance/revenue": {
      "get": {
        "operationId": "FinanceController_revenue",
        "parameters": [
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "channel",
                  "account",
                  "country",
                  "provider",
                  "plan"
                ]
              }
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "range": {
                    "from": "2026-08-24T23:00:00.000Z",
                    "to": "2026-09-24T22:59:59.999Z"
                  },
                  "series": [
                    {
                      "date": "2026-08-25T00:00:00.000Z",
                      "creditsConsumed": 0,
                      "subscriptionRevenueNgn": 0,
                      "cashCollectedNgn": 0,
                      "creditLiability": 0,
                      "mrrNgn": 0,
                      "providerCostNgn": null,
                      "grossProfitNgn": null
                    },
                    {
                      "date": "2026-08-26T00:00:00.000Z",
                      "creditsConsumed": 0,
                      "subscriptionRevenueNgn": 0,
                      "cashCollectedNgn": 0,
                      "creditLiability": 0,
                      "mrrNgn": 0,
                      "providerCostNgn": null,
                      "grossProfitNgn": null
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/admin/v1/finance/profit": {
      "get": {
        "operationId": "FinanceController_profit",
        "parameters": [
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "channel",
                  "account",
                  "country",
                  "provider",
                  "plan"
                ]
              }
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "range": {
                    "from": "2026-08-24T23:00:00.000Z",
                    "to": "2026-09-24T22:59:59.999Z"
                  },
                  "totals": {
                    "recognisedRevenueNgn": 403218,
                    "usageRevenueNgn": 53218,
                    "subscriptionRevenueNgn": 350000,
                    "providerCostNgn": 0,
                    "paymentFeesNgn": 0,
                    "grossProfitNgn": 403218,
                    "grossMargin": 1
                  },
                  "data": [
                    {
                      "channel": "EMAIL",
                      "revenueNgn": 2776.5,
                      "providerCostNgn": 0,
                      "contributionNgn": 2776.5,
                      "margin": 1,
                      "attempted": 1851,
                      "succeeded": 1791,
                      "failed": 60
                    },
                    {
                      "channel": "SMS",
                      "revenueNgn": 50441.5,
                      "providerCostNgn": 0,
                      "contributionNgn": 50441.5,
                      "margin": 1,
                      "attempted": 7461,
                      "succeeded": 6476,
                      "failed": 985
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/admin/v1/finance/cash": {
      "get": {
        "operationId": "FinanceController_cash",
        "parameters": [
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "channel",
                  "account",
                  "country",
                  "provider",
                  "plan"
                ]
              }
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "range": {
                    "from": "2026-08-24T23:00:00.000Z",
                    "to": "2026-09-24T22:59:59.999Z"
                  },
                  "cashCollectedNgn": 211500,
                  "grossCollectedNgn": 211500,
                  "refundedNgn": 0,
                  "paymentFeesNgn": null,
                  "successfulPayments": 45,
                  "failedPayments": 3,
                  "creditLiability": 157456
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/admin/v1/finance/cohorts": {
      "get": {
        "operationId": "FinanceController_cohorts",
        "parameters": [
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "channel",
                  "account",
                  "country",
                  "provider",
                  "plan"
                ]
              }
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "range": {
                    "from": "2026-08-24T23:00:00.000Z",
                    "to": "2026-09-24T22:59:59.999Z"
                  },
                  "activeAccounts": 5,
                  "arpaCredits": 10643.6,
                  "revenueChurn": null,
                  "churnedCredits": 0,
                  "priorPeriod": {
                    "from": "2026-07-24T23:00:00.001Z",
                    "to": "2026-08-24T23:00:00.000Z",
                    "revenueNgn": 0
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/admin/v1/finance/reconciliation": {
      "get": {
        "operationId": "FinanceController_reconciliationRuns",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "dcdf9f9a-04f2-4f00-bc05-f86947a97295",
                      "date": "2026-09-24T00:00:00.000Z",
                      "kind": "PAYSTACK_SETTLEMENT",
                      "status": "OK",
                      "expectedNgn": "22500",
                      "actualNgn": "22500",
                      "variancePct": "0",
                      "detail": {
                        "note": "Successful transactions against the cash rollup for the same Lagos day"
                      },
                      "createdAt": "2026-09-24T10:48:04.235Z"
                    },
                    {
                      "id": "5103593e-7315-499b-b28e-fa6652ca1ebb",
                      "date": "2026-09-23T00:00:00.000Z",
                      "kind": "PAYSTACK_SETTLEMENT",
                      "status": "OK",
                      "expectedNgn": "22500",
                      "actualNgn": "22500",
                      "variancePct": "0",
                      "detail": {
                        "note": "Successful transactions against the cash rollup for the same Lagos day"
                      },
                      "createdAt": "2026-09-23T12:02:33.787Z"
                    }
                  ],
                  "hasMore": false,
                  "nextCursor": null
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/admin/v1/exports": {
      "post": {
        "operationId": "FinanceController_createExport",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      },
      "get": {
        "operationId": "FinanceController_listExports",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "3f4753b9-ce5c-4df5-aef7-8cbb92824c39",
                      "requestedById": "5799da62-bd38-4ddf-8171-2b2592d79c0c",
                      "kind": "profit",
                      "params": {
                        "kind": "profit"
                      },
                      "containsPii": false,
                      "status": "FAILED",
                      "rowCount": null,
                      "objectKey": null,
                      "expiresAt": "2026-09-23T15:11:06.766Z",
                      "error": "The specified bucket does not exist",
                      "createdAt": "2026-09-22T15:11:06.767Z",
                      "completedAt": null
                    }
                  ],
                  "hasMore": false,
                  "nextCursor": null
                }
              }
            }
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/admin/v1/exports/{id}/download": {
      "get": {
        "operationId": "FinanceController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zapiack Admin Console API",
    "description": "Backend for the Zapiack Admin Console.\n\n**Authentication is two steps.** `POST /auth/login` returns `totp_required` and\nsets a session cookie that every guarded route rejects until\n`POST /auth/2fa/verify` succeeds. Send `credentials: \"include\"` on every request.\n\n**Writes need `x-csrf-token`**, echoing the value from `/auth/2fa/verify` or\n`/auth/me`. Some also need an `Idempotency-Key` header and a written `reason`;\neach operation says which.\n\n**Every list returns `{ data, hasMore, nextCursor }`.** Extra context sits beside\n`data`, never instead of it.\n\n**Money is a decimal string** (`\"125000.0000\"`), never a number. Dates are\nAfrica/Lagos calendar days. Unknown query or body fields are rejected with 400.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zpk_admin_at",
        "description": "Session cookie, set by the login flow. Not readable from JavaScript."
      }
    },
    "schemas": {
...
      "CreateExportDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "revenue",
              "profit",
              "cash",
              "ledger",
              "customers",
              "audit",
              "usage"
            ]
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          }
        },
        "required": [
          "kind"
        ]
      },
...
    }
  }
}